Add communication module
This module serves to define what messages are sent between two wallets.
