+++ packages/wallet/src/communication/index.ts
@@ -1,0 +1,8 @@
+export enum Strategy {
+  IndirectFunding = 'IndirectFunding',
+}
+
+export interface BaseProcessAction {
+  processId: string;
+  type: string;
+}
