+++ packages/wallet/src/redux/protocols/funding/index.ts
@@ -1,5 +1,2 @@
-export enum Strategy {
-  IndirectFunding = 'IndirectFunding',
-}
-
+export { Strategy } from '../../../communication';
 export { initialize, fundingReducer as reducer } from './reducer';
